requirements.txt
# Core
numpy>=1.26.0
scipy>=1.12.0
pandas>=2.2.0

# Retrieval
rank-bm25>=0.2.2
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0

# Recommendation baseline
recbole>=1.2.0

# LLM agent orchestration
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0

# LLM API
openai>=1.40.0
anthropic>=0.34.0

# Experiment tracking
mlflow>=2.14.0

# Visualization / demo
streamlit>=1.38.0
plotly>=5.22.0

# Config
pyyaml>=6.0.1
hydra-core>=1.3.2

# Development
pytest>=8.2.0
jupyter>=1.0.0
ipykernel>=6.29.0
